Adding AirCare
Thank you for publishing this great list! I'm one of the founders of [AirCare](https://air.care/) and we would be honored to be included. We've developed what we think is the easiest way to understand and compensate for the environmental impact of air travel. We have some really great features in the pipeline that we believe can have a huge positive impact.